chore: improved tests structure
