+++ tests/unit/HybridCSR_COO/HybridShard.cpp
@@ -4,7 +4,7 @@
 using namespace CinderPeak::PeakStore;
 using namespace std::chrono;
 
-class HybridCSRCOOTest : public ::testing::Test {
+class HybridStorageShardTest : public ::testing::Test {
 protected:
   void SetUp() override {
     graph = std::make_unique<HybridCSR_COO<int, int>>();
@@ -41,7 +41,7 @@ protected:
   std::unique_ptr<HybridCSR_COO<std::string, double>> string_graph;
 };
 
-TEST_F(HybridCSRCOOTest, EmptyGraphOperations) {
+TEST_F(HybridStorageShardTest, EmptyGraphOperations) {
   auto [weight, status] = graph->impl_getEdge(1, 2);
   EXPECT_FALSE(status.isOK());
 
@@ -49,7 +49,7 @@ TEST_F(HybridCSRCOOTest, EmptyGraphOperations) {
   EXPECT_FALSE(graph->impl_doesEdgeExist(1, 2, 10));
 }
 
-TEST_F(HybridCSRCOOTest, SingleVertexOperations) {
+TEST_F(HybridStorageShardTest, SingleVertexOperations) {
   auto status = graph->impl_addVertex(42);
   EXPECT_TRUE(status.isOK());
 
@@ -64,7 +64,7 @@ TEST_F(HybridCSRCOOTest, SingleVertexOperations) {
   EXPECT_EQ(weight, 100);
 }
 
-TEST_F(HybridCSRCOOTest, CheckVertexExistence) {
+TEST_F(HybridStorageShardTest, CheckVertexExistence) {
   auto status = graph->impl_addVertex(40);
   EXPECT_TRUE(status.isOK());
   auto status1 = graph->impl_addVertex(49);
@@ -76,7 +76,7 @@ TEST_F(HybridCSRCOOTest, CheckVertexExistence) {
   EXPECT_FALSE(graph->impl_hasVertex(404));
 }
 
-TEST_F(HybridCSRCOOTest, VertexAdditionSequential) {
+TEST_F(HybridStorageShardTest, VertexAdditionSequential) {
   std::vector<int> vertices = {1, 5, 3, 9, 2, 7};
 
   for (int v : vertices) {
@@ -90,7 +90,7 @@ TEST_F(HybridCSRCOOTest, VertexAdditionSequential) {
   }
 }
 
-TEST_F(HybridCSRCOOTest, EdgeAdditionBasic) {
+TEST_F(HybridStorageShardTest, EdgeAdditionBasic) {
   std::vector<int> vertices = {1, 2, 3, 4, 5};
   for (int v : vertices) {
     graph->impl_addVertex(v);
@@ -111,7 +111,7 @@ TEST_F(HybridCSRCOOTest, EdgeAdditionBasic) {
   EXPECT_EQ(w2, 25);
 }
 
-TEST_F(HybridCSRCOOTest, EdgeWeightUpdation) {
+TEST_F(HybridStorageShardTest, EdgeWeightUpdation) {
   std::vector<int> vertices = {1, 2, 3, 4, 5};
   for (int v : vertices) {
     graph->impl_addVertex(v);
@@ -134,7 +134,7 @@ TEST_F(HybridCSRCOOTest, EdgeWeightUpdation) {
   EXPECT_EQ(w2, 10);
 }
 
-TEST_F(HybridCSRCOOTest, EdgeAdditionWithNonExistentVertices) {
+TEST_F(HybridStorageShardTest, EdgeAdditionWithNonExistentVertices) {
   graph->impl_addVertex(1);
 
   auto status1 = graph->impl_addEdge(99, 1, 10);
@@ -147,7 +147,7 @@ TEST_F(HybridCSRCOOTest, EdgeAdditionWithNonExistentVertices) {
   EXPECT_FALSE(status3.isOK());
 }
 
-TEST_F(HybridCSRCOOTest, ClearEdges) {
+TEST_F(HybridStorageShardTest, ClearEdges) {
   std::vector<int> vertices = {1, 2, 3, 4, 5};
   for (int v : vertices) {
     graph->impl_addVertex(v);
@@ -172,7 +172,7 @@ TEST_F(HybridCSRCOOTest, ClearEdges) {
   EXPECT_FALSE(graph->impl_addVertex(4).isOK());
 }
 
-TEST_F(HybridCSRCOOTest, EdgeRetrievalAdvanced) {
+TEST_F(HybridStorageShardTest, EdgeRetrievalAdvanced) {
   for (int i = 1; i <= 5; ++i) {
     graph->impl_addVertex(i);
   }
@@ -197,7 +197,7 @@ TEST_F(HybridCSRCOOTest, EdgeRetrievalAdvanced) {
   EXPECT_FALSE(graph->impl_doesEdgeExist(5, 1));
 }
 
-TEST_F(HybridCSRCOOTest, COOBufferPriority) {
+TEST_F(HybridStorageShardTest, COOBufferPriority) {
   graph->impl_addVertex(1);
   graph->impl_addVertex(2);
 
@@ -215,7 +215,7 @@ TEST_F(HybridCSRCOOTest, COOBufferPriority) {
   EXPECT_EQ(weight, 999);
 }
 
-TEST_F(HybridCSRCOOTest, COOBufferOverwrite) {
+TEST_F(HybridStorageShardTest, COOBufferOverwrite) {
   graph->impl_addVertex(1);
   graph->impl_addVertex(2);
 
@@ -228,7 +228,7 @@ TEST_F(HybridCSRCOOTest, COOBufferOverwrite) {
   EXPECT_EQ(weight, 30);
 }
 
-TEST_F(HybridCSRCOOTest, StringVertexOperations) {
+TEST_F(HybridStorageShardTest, StringVertexOperations) {
   auto status1 = string_graph->impl_addVertex("alice");
   auto status2 = string_graph->impl_addVertex("bob");
   auto status3 = string_graph->impl_addVertex("charlie");
@@ -253,7 +253,7 @@ TEST_F(HybridCSRCOOTest, StringVertexOperations) {
   EXPECT_FALSE(s3.isOK());
 }
 
-TEST_F(HybridCSRCOOTest, NegativeWeights) {
+TEST_F(HybridStorageShardTest, NegativeWeights) {
   graph->impl_addVertex(1);
   graph->impl_addVertex(2);
 
@@ -267,7 +267,7 @@ TEST_F(HybridCSRCOOTest, NegativeWeights) {
   EXPECT_FALSE(graph->impl_doesEdgeExist(1, 2, 100));
 }
 
-TEST_F(HybridCSRCOOTest, ZeroWeights) {
+TEST_F(HybridStorageShardTest, ZeroWeights) {
   graph->impl_addVertex(1);
   graph->impl_addVertex(2);
 
@@ -278,7 +278,7 @@ TEST_F(HybridCSRCOOTest, ZeroWeights) {
   EXPECT_EQ(weight, 0);
 }
 
-TEST_F(HybridCSRCOOTest, LargeVertexIDs) {
+TEST_F(HybridStorageShardTest, LargeVertexIDs) {
   int large_id1 = 1000000;
   int large_id2 = 2000000;
 
@@ -292,7 +292,7 @@ TEST_F(HybridCSRCOOTest, LargeVertexIDs) {
   EXPECT_EQ(weight, 42);
 }
 
-TEST_F(HybridCSRCOOTest, PopulateFromAdjacencyList) {
+TEST_F(HybridStorageShardTest, PopulateFromAdjacencyList) {
   std::unordered_map<int, std::vector<std::pair<int, int>>,
                      CinderPeak::VertexHasher<int>>
       adj_list;
@@ -321,7 +321,7 @@ TEST_F(HybridCSRCOOTest, PopulateFromAdjacencyList) {
   EXPECT_FALSE(s21.isOK());
 }
 
-TEST_F(HybridCSRCOOTest, PopulateFromEmptyAdjacencyList) {
+TEST_F(HybridStorageShardTest, PopulateFromEmptyAdjacencyList) {
   std::unordered_map<int, std::vector<std::pair<int, int>>,
                      CinderPeak::VertexHasher<int>>
       empty_adj_list;
@@ -332,7 +332,7 @@ TEST_F(HybridCSRCOOTest, PopulateFromEmptyAdjacencyList) {
   EXPECT_FALSE(status.isOK());
 }
 
-TEST_F(HybridCSRCOOTest, ManyVerticesNoEdges) {
+TEST_F(HybridStorageShardTest, ManyVerticesNoEdges) {
   const int NUM_VERTICES = 10000;
 
   for (int i = 0; i < NUM_VERTICES; ++i) {
@@ -344,7 +344,7 @@ TEST_F(HybridCSRCOOTest, ManyVerticesNoEdges) {
   EXPECT_FALSE(status.isOK());
 }
 
-TEST_F(HybridCSRCOOTest, HighDegreeVertex) {
+TEST_F(HybridStorageShardTest, HighDegreeVertex) {
   const int HUB_VERTEX = 0;
   const int NUM_NEIGHBORS = 1000;
 
@@ -367,7 +367,7 @@ TEST_F(HybridCSRCOOTest, HighDegreeVertex) {
   }
 }
 
-TEST_F(HybridCSRCOOTest, DenseGraph) {
+TEST_F(HybridStorageShardTest, DenseGraph) {
   const int NUM_VERTICES = 100;
 
   for (int i = 0; i < NUM_VERTICES; ++i) {
